pyproject.toml
[project]
name = "fapi-db-tmpl"
version = "0.1.0"
description = "A production-ready FastAPI project template with database."
authors = [{ name = "akitorahayashi" }]
requires-python = ">=3.12,<4.0"
dependencies = [
    "alembic>=1.13.1,<2.0.0",
    "fastapi>=0.116.1,<0.117.0",
    "httpx>=0.27.0,<0.28.0",
    "psycopg[binary]>=3.1,<4.0",
    "pydantic-settings>=2.4.0,<3.0.0",
    "sqlalchemy>=2.0.29,<3.0.0",
    "uvicorn>=0.35.0,<0.36.0",
    "uvloop>=0.19.0,<0.20.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=0.23.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "black>=25.1.0",
    "ruff>=0.12.10",
    "testcontainers[compose]>=4.8.0,<5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fapi_db_tmpl"]

[tool.pytest.ini_options]
python_files = "test_*.py"
asyncio_mode = "auto"
addopts = "-v -s --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.black]
target-version = ['py312']
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.uv-cache
    | \.venv
    | venv
    | env
    | build
    | dist
    | reference
    | .*/(bin|lib|share|etc)
    | .*/pyvenv\.cfg
)/
'''

[tool.ruff]
target-version = "py312"
exclude = [
    "__pycache__",
    "migrations",
    "db.sqlite3",
    "*.pyc",
    ".uv-cache",
    ".venv",
    "venv",
    "env",
    "reference",
    "*/bin/**",
    "*/lib/**",
    "*/share/**",
    "*/etc/**",
    "*/pyvenv.cfg"
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"] # E501 (line too long) is ignored

[tool.alembic]
script_location = "alembic"
prepend_sys_path = "."
path_separator = "os"
